Separate named pipes for tray icon and cred provider
