+++ Lanpartyseating.Desktop/Business/ICredentialProviderPipeService.cs
@@ -1,0 +1,6 @@
+namespace Lanpartyseating.Desktop.Business;
+
+public interface ICredentialProviderPipeService
+{
+    Task TriggerLoginAsync(string username, string password, string? domain = null);
+}
