--- Lanpartyseating.Desktop/Business/INamedPipeServerService.cs
@@ -1,9 +1,0 @@
-﻿using Lanpartyseating.Desktop.Abstractions;
-
-namespace Lanpartyseating.Desktop;
-
-public interface INamedPipeServerService
-{
-    Task SendMessageAsync<T>(T message, CancellationToken cancellationToken) where T : BaseMessage;
-    Task TriggerLoginAsync(string username, string password, string? domain = null);
-}
